Extract ModbusTCP decorator
