+++ src/deye_modbus.py
@@ -19,7 +19,6 @@ import logging
 
 import libscrc
 
-from deye_config import DeyeConfig
 from deye_connector import DeyeConnector
 
 
@@ -28,9 +27,8 @@ class DeyeModbus:
     Inspired by https://github.com/jlopez77/DeyeInverter
     """
 
-    def __init__(self, config: DeyeConfig, connector: DeyeConnector):
+    def __init__(self, connector: DeyeConnector):
         self.__log = logging.getLogger(DeyeModbus.__name__)
-        self.config = config.logger
         self.connector = connector
 
     def read_registers(self, first_reg: int, last_reg: int) -> dict[int, bytearray]:
@@ -45,9 +43,10 @@ class DeyeModbus:
             and register value is the map value
         """
         modbus_frame = self.__build_modbus_read_holding_registers_request_frame(first_reg, last_reg)
-        req_frame = self.__build_request_frame(modbus_frame)
-        resp_frame = self.connector.send_request(req_frame)
-        modbus_resp_frame = self.__extract_modbus_response_frame(resp_frame)
+        modbus_crc = bytearray.fromhex("{:04x}".format(libscrc.modbus(modbus_frame)))
+        modbus_crc.reverse()
+
+        modbus_resp_frame = self.connector.send_request(modbus_frame + modbus_crc)
         if modbus_resp_frame is None:
             return {}
         return self.__parse_modbus_read_holding_registers_response(modbus_resp_frame, first_reg, last_reg)
@@ -77,64 +76,13 @@ class DeyeModbus:
         """
 
         modbus_frame = self.__build_modbus_write_holding_register_request_frame(reg_address, reg_values)
-        req_frame = self.__build_request_frame(modbus_frame)
-        resp_frame = self.connector.send_request(req_frame)
-        modbus_resp_frame = self.__extract_modbus_response_frame(resp_frame)
+        modbus_crc = bytearray.fromhex("{:04x}".format(libscrc.modbus(modbus_frame)))
+        modbus_crc.reverse()
+
+        modbus_resp_frame = self.connector.send_request(modbus_frame + modbus_crc)
         if modbus_resp_frame is None:
             return False
         return self.__parse_modbus_write_holding_register_response(modbus_resp_frame, reg_address, reg_values)
-
-    def __build_request_frame(self, modbus_frame) -> bytearray:
-        start = bytearray.fromhex("A5")  # start
-        length = (15 + len(modbus_frame) + 2).to_bytes(2, "little")  # datalength
-        controlcode = bytearray.fromhex("1045")  # controlCode
-        inverter_sn_prefix = bytearray.fromhex("0000")  # serial
-        datafield = bytearray.fromhex("020000000000000000000000000000")
-        modbus_crc = bytearray.fromhex("{:04x}".format(libscrc.modbus(modbus_frame)))
-        modbus_crc.reverse()
-        checksum = bytearray.fromhex("00")  # checksum placeholder for outer frame
-        end_code = bytearray.fromhex("15")
-        inverter_sn = bytearray.fromhex("{:10x}".format(self.config.serial_number))
-        inverter_sn.reverse()
-        frame = (
-            start
-            + length
-            + controlcode
-            + inverter_sn_prefix
-            + inverter_sn
-            + datafield
-            + modbus_frame
-            + modbus_crc
-            + checksum
-            + end_code
-        )
-
-        checksum = 0
-        for i in range(1, len(frame) - 2, 1):
-            checksum += frame[i] & 255
-        frame[len(frame) - 2] = int((checksum & 255))
-
-        return frame
-
-    def __extract_modbus_response_frame(self, frame: bytes | None) -> bytes | None:
-        # 29 - outer frame, 2 - modbus addr and command, 2 - modbus crc
-        if not frame:
-            # Error was already logged in `send_request()` function
-            return None
-        if len(frame) == 29:
-            self.__parse_response_error_code(frame)
-            return None
-        if len(frame) < (29 + 4):
-            self.__log.error("Response frame is too short")
-            return None
-        if frame[0] != 0xA5:
-            self.__log.error("Response frame has invalid starting byte")
-            return None
-        if frame[-1] != 0x15:
-            self.__log.error("Response frame has invalid ending byte")
-            return None
-
-        return frame[25:-2]
 
     def __build_modbus_read_holding_registers_request_frame(self, first_reg: int, last_reg: int) -> bytearray:
         reg_count = last_reg - first_reg + 1
@@ -201,13 +149,3 @@ class DeyeModbus:
             )
             return False
         return True
-
-    def __parse_response_error_code(self, frame: bytes) -> None:
-        error_frame = frame[25:-2]
-        error_code = error_frame[0]
-        if error_code == 0x05:
-            self.__log.error("Modbus device address does not match.")
-        elif error_code == 0x06:
-            self.__log.error("Logger Serial Number does not match. Check your configuration file.")
-        else:
-            self.__log.error("Unknown response error code. Error frame: %s", error_frame.hex())
