+++ tests/deye_modbus_test.py
@@ -20,6 +20,7 @@ from unittest.mock import patch
 
 from deye_config import DeyeConfig, DeyeLoggerConfig
 from deye_modbus import DeyeModbus
+from deye_modbus_tcp import DeyeModbusTcp
 
 
 class DeyeModbusTest(unittest.TestCase):
@@ -29,7 +30,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_read_register_0x01(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a5000000000000000000000000000000000000000000000000010301000ac84300000015"
         )
@@ -45,7 +46,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_read_registers_0x02_0x03(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a5000000000000000000000000000000000000000000000000010302000a000b13f600000015"
         )
@@ -68,7 +69,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_write_register_0x12_to_0xa3d4(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a5000000000000000000000000000000000000000000000000" + "011000120001a1cc" + "0015"
         )
@@ -85,7 +86,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_read_register_SUN_10K_SG04LP3_EU_part1(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a53b0010150007482ee38d020121d0060091010000403e486301032800ffffff160a12162420ffffffffffffffffffffffffffffffffffff0001ffff0001ffff000003e81fa45115"
         )
@@ -103,7 +104,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_read_register_SUN_10K_SG04LP3_EU_part2(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a5330010150008482ee38d020122d0060091010000403e486301032000010000ffffffffffff0001ffffffffffffffffffff0000ffff0011ffffffff3a005715"
         )
@@ -121,7 +122,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_incorrect_inverter_serial_number(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a51000101500c9c22576f80201032d0000790800007106d6630600bd15"
         )
@@ -139,7 +140,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_incorrect_modbus_address(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a51000101500c9c22576f80201032d0000790800007106d6630500bd15"
         )
@@ -155,7 +156,7 @@ class DeyeModbusTest(unittest.TestCase):
     @patch("deye_connector.DeyeConnector")
     def test_unknown_error_code(self, connector):
         # given
-        sut = DeyeModbus(self.config, connector)
+        sut = DeyeModbus(DeyeModbusTcp(self.config, connector))
         connector.send_request.return_value = bytearray.fromhex(
             "a51000101500c9c22576f80201032d0000790800007106d6630100bd15"
         )
